feat: apply i18n to playground
