+++ app/types/dictionaries/index.ts
@@ -1,0 +1,1 @@
+export * from './playground';
