+++ app/types/dictionaries/playground.ts
@@ -1,0 +1,22 @@
+export type PlaygroundDictionary = {
+  playground: {
+    title: string;
+    description: string;
+    string: {
+      placeholder: string;
+    };
+    array: {
+      placeholder: string;
+    };
+    separator: {
+      description: string;
+    };
+    output: {
+      title: string;
+    };
+    button: {
+      run: string;
+      back: string;
+    };
+  };
+};
